+++ src/lib/Containers/WorksForYou.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { createPaginationContainer, graphql } from "react-relay"
+import { ConnectionData, createPaginationContainer, graphql, RelayPaginationProp } from "react-relay"
 
 import {
   LayoutChangeEvent,
@@ -22,19 +22,18 @@ GenericGrid
 import ZeroState from "lib/Components/States/ZeroState"
 import SerifText from "lib/Components/Text/Serif"
 import Notification from "lib/Components/WorksForYou/Notification"
+import { isCloseToBottom } from "lib/utils/isCloseToBottom"
 
 import colors from "lib/data/colors"
 
 const PageSize = 10
 
 interface Props extends RelayProps {
-  relay: any
+  relay?: RelayPaginationProp
 }
 
 interface State {
   dataSource: ListViewDataSource | null
-  fetchingNextPage: boolean
-  completed: boolean
 }
 
 export class WorksForYou extends React.Component<Props, State> {
@@ -42,7 +41,7 @@ export class WorksForYou extends React.Component<Props, State> {
   scrollView?: ScrollView | any
   currentScrollOffset?: number = 0
 
-  constructor(props) {
+  constructor(props: Props) {
     super(props)
 
     const notifications = this.props.viewer.me.notifications.edges.map(edge => edge.node)
@@ -56,8 +55,6 @@ export class WorksForYou extends React.Component<Props, State> {
 
     this.state = {
       dataSource,
-      completed: false,
-      fetchingNextPage: false,
     }
   }
 
@@ -94,11 +91,10 @@ export class WorksForYou extends React.Component<Props, State> {
     }
   }
 
-  fetchNextPage() {
-    if (this.state.fetchingNextPage || this.state.completed) {
+  fetchNextPage = () => {
+    if (!this.props.relay.hasMore() || this.props.relay.isLoading()) {
       return
     }
-    this.setState({ fetchingNextPage: true })
     this.props.relay.loadMore(PageSize, error => {
       const notifications = this.props.viewer.me.notifications.edges.map(edge => edge.node)
 
@@ -108,17 +104,9 @@ export class WorksForYou extends React.Component<Props, State> {
       }
 
       this.setState({
-        fetchingNextPage: false,
         dataSource: this.state.dataSource.cloneWithRows(notifications),
       })
-      if (!this.props.viewer.me.notifications.pageInfo.hasNextPage) {
-        this.setState({ completed: true })
-      }
     })
-  }
-
-  componentDidUpdate() {
-    this.scrollView.scrollTo({ y: this.currentScrollOffset + 1, animated: false })
   }
 
   render() {
@@ -130,7 +118,7 @@ export class WorksForYou extends React.Component<Props, State> {
     return (
       <ScrollView
         contentContainerStyle={hasNotifications ? {} : styles.container}
-        onScroll={event => (this.currentScrollOffset = event.nativeEvent.contentOffset.y)}
+        onScroll={isCloseToBottom(this.fetchNextPage)}
         scrollEventThrottle={100}
         ref={scrollView => (this.scrollView = scrollView)}
       >
@@ -148,7 +136,6 @@ export class WorksForYou extends React.Component<Props, State> {
         renderRow={data => <Notification notification={data} />}
         renderSeparator={(sectionID, rowID) =>
           <View key={`${sectionID}-${rowID}`} style={styles.separator} /> as React.ReactElement<{}>}
-        onEndReached={() => this.fetchNextPage()}
         scrollEnabled={false}
       />
     )
@@ -227,7 +214,7 @@ const WorksForYouContainer = createPaginationContainer(
   {
     direction: "forward",
     getConnectionFromProps(props) {
-      return props.viewer.me.notifications
+      return props.viewer.me.notifications as ConnectionData
     },
     getFragmentVariables(prevVars, totalCount) {
       return {